refactor endpoint services
